[level 0] Title: 몫 구하기, Time: 0.01 ms, Memory: 4.14 MB -BaekjoonHub

+++ 프로그래머스/lv0/120805. 몫 구하기/몫 구하기.cpp
@@ -4,5 +4,5 @@
 using namespace std;
 
 int solution(int num1, int num2) {
-    return int(num1/num2);
+    return num1/num2;
 }
